requirements.txt
langchain
streamlit 
pdf2image
pdfminer
pdfminer.six
pillow-heif
opencv-python
unstructured_inference 
pytesseract
tesseract-ocr
unstructured_pytesseract
pikepdf 
pypdf
docx
python-docx
chromadb
langchain_google_genai
unstructed
